cache database on client ### motivation: frequent reloads caused by remix actions currently hit the fs too much in workspace loader and cause delays proportional to number of requests
### idea:
recreate the client copy of the database previous in redux. This will extend the lifespan of the data beyond the current render.

### options:
- exact copy of the whole database
- copy of this workspace data
- copy of the sidebar data
<!--
Please open an [Issue](https://github.com/kong/insomnia/issues/new) first to discuss new
features or non-trivial changes. Please provide as much detail as possible on the change as
possible including general description, implementation details, potential shortcomings, etc.

If this PR closes an issue, please mention "Closes #XX" where #XX is the issue number.

If this PR fixes a bug or regression, please make sure to add a test.
-->
